+++ NEWS.md
@@ -13,6 +13,8 @@ argument is specified as a single function rather than a named list. Thanks to @
 * For `ppc_violin_grouped` there is now the option of showing `y` as a violin, 
 points, or both. Thanks to @silberzwiebel. (#74)
 * New color scheme `"darkgray"`.
+* `color_scheme_get` now has an optional argument `i` for selecting only a 
+subset of the colors.
 
 #### New functions
 * `mcmc_pairs`, which is essentially a ggplot2+grid implementation of rstan's